refactor(spa): extract slugFromTitle to shared utility + migrate tests (T6) [KAN-118]
Deduplicate slugFromTitle from GeneratorComponent and RecipeDetailComponent
into src/utils/slug.ts with Unicode property class regex. Slug parity tests
(15 cases) moved to src/utils/slug.test.ts. Component test file cleaned up
to only contain modal-specific tests.
