+++ src/utils/slug.ts
@@ -1,0 +1,8 @@
+export function slugFromTitle(title: string): string {
+  return title
+    .normalize('NFKD')
+    .replace(/[^\p{ASCII}]/gu, '')
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, '-')
+    .replace(/^-+|-+$/g, '');
+}
